requirements.txt
smolagents ==1.2.2
huggingface_hub
litellm
markdownify
duckduckgo-search
transformers
beautifulsoup4
requests
langchain
langchain-community
sentence-transformers
python-dotenv
rank_bm25
faiss-cpu